PlutoVIP was broken.

+++ src/PlutoVIP/PlutoVIPRequests.cpp
@@ -29,7 +29,7 @@
 #include "VA_SendFileToPhone.h"
 #include "VA_SendMenuToPhone.h"
 
-RA_Request *RA_Processor::BuildRequestFromData(long dwSize, const char *pcData, unsigned long dwRequestID)
+RA_Request *RA_Processor::BuildRequestFromData(uint32_t dwSize, const char *pcData, uint32_t dwRequestID)
 {
 	RA_Request *pRequest=NULL;
 	switch( dwRequestID )
@@ -69,7 +69,7 @@ RA_Request *RA_Processor::BuildRequestFromData(long dwSize, const char *pcData, 
 	return pRequest;
 }
 
-RA_Action *RA_Processor::BuildActionFromData( long dwSize, const char *pcData, unsigned long dwActionID)
+RA_Action *RA_Processor::BuildActionFromData( uint32_t dwSize, const char *pcData, uint32_t dwActionID)
 {
 	RA_Action *pRA_Action=NULL;
 	switch( dwActionID )
